requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.8
dnspython==2.6.1
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
identify==2.6.0
idna==3.7
kaleido==0.2.1
kiwisolver==1.4.5
matplotlib==3.9.1
multidict==6.0.5
nodeenv==1.9.1
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
pre-commit==3.8.0
pycountry==24.6.1
pymongo==4.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
schedule==1.2.2
six==1.16.0
tenacity==9.0.0
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
yarl==1.9.4
